refactor: extract event-cardinality bootverify checks
